Enhance difficulty select UI, font, and data management
- Add per-difficulty custom fonts, colors, and scaling via `description.json`
- Refactor font creation, copying, and cleanup for safety and flexibility
- Highlight selected difficulty and "More or Less" text in gold
- Track used image indexes to prevent duplicate JSON/video entries
- Update storage handling to avoid duplicates and update in-use status

The change in the Python side of this fixes the bug where the sub/view data
and the thumbnail shown do not
- Add new font assets and update all test images
- Fix typos in `description.json` and improve float parsing
- Various UI/UX improvements, bugfixes, and logging enhancements

+++ src/globals.h
@@ -17,7 +17,7 @@ Screen* screen;
 typedef struct fonts {
 	TTF_Font** fonts;
 	int fontIndex;
-	float fontSize;
+	int fontSize;
 } Fonts;
 
 typedef struct ytNode {
